move catalog product to dedicated folder add localization to catalog product
