+++ src/modules/catalog-product/components/.gitkeep:Zone.Identifier
@@ -1,0 +1,2 @@
+[ZoneTransfer]
+ZoneId=3
