+++ src/modules/catalog-product/components/ProductVariantPill.vue
