+++ src/modules/catalog-product/interfaces/index.ts
@@ -2,6 +2,38 @@ export interface ICategory {
   id: string;
   category: string;
   description?: string;
+  pictureUrl?: string;
+}
+
+export interface CategoryPayload {
+  imagePreview?: string;
+  imageFile?: File;
+  category: string;
+  description?: string;
+}
+
+export interface ICategoryFormData {
+  imagePreview?: string;
+  imageFile?: File;
+  name: string;
+  description: string;
+}
+
+export interface ICategoryResponse {
+  categories: ICategory[];
+  lastPage: number;
+}
+
+export interface ICategoryAddResponse {
+  statusCode: number;
+  message: string;
+  data?: ICategory;
+}
+
+export interface ICategory {
+  id: string;
+  category: string;
+  description?: string;
 }
 
 export interface IVariant {
@@ -52,4 +84,10 @@ export interface CreateProductPayload {
 export interface IProductResponse {
   products: IProduct[];
   lastPage: number;
+}
+
+
+export interface CreateCategoryPayload {
+  name: string;
+  price?: number;
 }
