+++ src/modules/catalog-product/interfaces/index.ts:Zone.Identifier
@@ -1,0 +1,2 @@
+[ZoneTransfer]
+ZoneId=3
