+++ src/modules/catalog-product/views/EditProductUI.vue
@@ -1,29 +1,84 @@
 <script setup>
-import { useProductService } from '@/modules/catalog/services/Product/ProductServices';
-import { useCategoryService } from '../../services/Category/CategoryService';
+import { useProductService } from '../services/catalog-product.service';
+import { useCategoryService } from '@/modules/catalog/services/Category/CategoryService';
 import excludeSVG from '@/app/assets/icons/exclude.svg';
-import closeRedSVG from '@/app/assets/icons/close-red.svg';
+import deleteSVG from '@/app/assets/icons/delete.svg';
 import plusLineSVG from '@/app/assets/icons/plus-line.svg';
 import chevronDownSVG from '@/app/assets/icons/chevron-down.svg';
 import imageSVG from '@/app/assets/icons/image.svg';
+import confirmationSVG from '@/app/assets/icons/confirmation.svg';
+import closeRedSVG from '@/app/assets/icons/close-red.svg';
+import deletePolygonSVG from '@/app/assets/icons/delete-polygon.svg';
+
+const route = useRoute();
 
 const { getAllCategories } = useCategoryService();
-const { createProduct, product_formData, product_formValidations } = useProductService();
-const discount_unit = ref('Rp');
+const { getProductById, updateProduct, deleteProduct, product_formData, product_formValidations } =
+  useProductService();
+
+const toggleVariant = ref(false);
+const categories = ref([]);
+const productID = ref(route.params.id);
 function clearForm() {
   product_formData.name = '';
   product_formData.price = 0;
   product_formData.isDiscount = false;
   product_formData.discount_value = 0;
+  product_formData.is_percent = false;
   product_formData.discount_price = 0;
   product_formData.variants = [];
   product_formData.categories = [];
   product_formData.imagePreview = '';
 
+  toggleVariant.value = false;
   product_formValidations.value.$reset();
 }
 
+const loadCategories = async () => {
+  try {
+    const response = await getAllCategories(1, 100, '');
+    categories.value = response.categories;
+  } catch (error) {
+    console.error('Failed to load categories:', error);
+  }
+};
+
+const discount_unit = ref('Rp');
+
+const loadProduct = async () => {
+  try {
+    const response = await getProductById(route.params.id);
+
+    product_formData.name = response.name;
+    product_formData.price = response.price;
+    product_formData.discount_price = response.discountPrice;
+    product_formData.variants = response.variantHasProducts;
+    product_formData.categories = response.categoriesHasProducts;
+    product_formData.is_percent = response.isPercent;
+    console.log("🚀 ~ loadProduct ~ response.isPercent:", response.isPercent)
+    product_formData.imagePreview = response.picture_url;
+
+    product_formData.isDiscount = product_formData.price !== product_formData.discount_price;
+
+    if (product_formData.isDiscount) {
+      if (response.isPercent) {
+        product_formData.discount_value = 100 - (response.discountPrice / response.price) * 100;
+        discount_unit.value = '%';
+      } else {
+        product_formData.discount_value = response.price - response.discountPrice;
+        discount_unit.value = 'Rp';
+      }
+    }
+    if (product_formData.variants.length > 0) {
+      toggleVariant.value = true;
+    }
+  } catch (error) {
+    console.error(error);
+  }
+};
+
 const fileInput = ref(null);
+
 const triggerFileInput = () => {
   fileInput.value?.click();
 };
@@ -41,29 +96,28 @@ const handleImageUpload = event => {
   }
 };
 
-const handleCreateProduct = async () => {
-  product_formValidations.value.$touch();
-  if (product_formValidations.value.$invalid) return;
+const isDeleteOpen = ref(false);
 
-  if (!product_formData.isDiscount) {
-    // console.log('no discount');
-    product_formData.discount_price = product_formData.price;
-  }
-
+const handleDelete = async () => {
   try {
-    await createProduct(product_formData);
-    clearForm();
-    product_formValidations.value.$reset();
+    await deleteProduct(productID.value);
+    isDeleteOpen.value = false;
     hasConfirmedLeave = true;
     router.push({ name: 'catalog.products.index' });
   } catch (error) {
-    console.error(error);
+    console.error('Failed to delete product:', error);
   }
 };
 
-const categories = ref([]);
-
-const toggleVariant = ref(false);
+const handleUpdateProduct = async () => {
+  try {
+    await updateProduct(productID.value, product_formData);
+  } catch (error) {
+    console.error(error);
+  } finally {
+    clearForm();
+  }
+};
 
 const addVariant = () => {
   product_formData.variants.push({
@@ -98,19 +152,6 @@ const isLeavingModal = ref(false);
 
 const router = useRouter();
 
-const loadCategories = async () => {
-  try {
-    const response = await getAllCategories(1, 100, '');
-    categories.value = response.categories;
-  } catch (error) {
-    console.error('Failed to load categories:', error);
-  }
-};
-
-onMounted(async () => {
-  loadCategories();
-});
-
 let hasConfirmedLeave = false;
 
 const confirmLeave = () => {
@@ -123,6 +164,11 @@ const confirmLeave = () => {
     router.push(targetRoute);
   }
 };
+
+onMounted(async () => {
+  loadCategories();
+  loadProduct();
+});
 
 const cancelLeave = () => {
   isLeavingModal.value = false;
@@ -145,33 +191,48 @@ onBeforeRouteLeave((to, from, next) => {
   }
 });
 
-watch(product_formData, () => {
-  calculateDiscount();
-});
+// Modal state
+const isUpdateModal = ref(false);
+
+// Your action to call when confirming
+const confirmUpdate = async () => {
+  try {
+    // Example: updateProduct() or any async call
+    await handleUpdateProduct(); // Make sure this is defined/imported
+
+    product_formValidations.value.$reset();
+    hasConfirmedLeave = true;
+    router.push({ name: 'catalog.products.index' });
+    // Optionally show success feedback/toast here
+
+    isUpdateModal.value = false;
+  } catch (error) {
+    console.error('Failed to update product:', error);
+    // Optionally show error feedback
+  }
+};
+
+// Cancel just closes the modal
+const cancelUpdate = () => {
+  isUpdateModal.value = false;
+};
 </script>
 
 <template>
   <div class="container mx-auto">
     <div class="flex flex-col gap-4">
-      <!-- {{ product_formData }} -->
-      <!-- <br />
-      {{ product_formValidations }}
-      <br />
-      {{ product_formValidations.$invalid }} -->
       <h1 class="text-2xl font-bold">Products Detail</h1>
       <h2 class="text-xl font-semibold">Product Information</h2>
-      <form class="flex flex-col items-center justify-center" @submit.prevent="handleCreateProduct">
+
+      <form class="flex flex-col items-center justify-center" @submit.prevent="isUpdateModal = true">
         <p>Photo (Optional)</p>
         <img
           class="rounded-lg mt-2 w-64 h-64 object-cover"
           :src="product_formData.imagePreview || 'https://placehold.co/250'"
           alt="Photo"
         />
-
-        <!-- Hidden File Input -->
         <input ref="fileInput" type="file" accept="image/*" class="hidden" @change="handleImageUpload" />
 
-        <!-- PrimeVue Button as file selector -->
         <PrimeVueButton
           label="Select Image"
           class="mt-4 shadow-xs hover:bg-transparent rounded-xl px-8 py-2 text-primary border-primary border-2"
@@ -196,7 +257,7 @@ watch(product_formData, () => {
               v-model="product_formData.name"
               name="name"
               type="text"
-              :class="{ ...classes }"
+              :class="classes ? '' : ''"
               fluid
               class="border shadow-xs border-grayscale-30 rounded-lg p-2 w-full"
               v-on="useListenerForm(product_formValidations, 'name')"
@@ -221,7 +282,7 @@ watch(product_formData, () => {
                 filter
                 placeholder="Select"
                 class="w-full text-primary"
-                :class="{ ...classes }"
+                :class="classes ? '' : ''"
                 v-on="useListenerForm(product_formValidations, 'categories')"
               >
                 <template #dropdownicon>
@@ -248,7 +309,7 @@ watch(product_formData, () => {
                 name="price"
                 fluid
                 class="border shadow-xs border-grayscale-30 rounded-lg"
-                :class="{ ...classes }"
+                :class="classes ? '' : ''"
                 v-on="useListenerForm(product_formValidations, 'price')"
                 @change="calculateDiscount"
               />
@@ -280,7 +341,7 @@ watch(product_formData, () => {
                     :prefix="product_formData.is_percent === false ? 'Rp ' : ''"
                     :suffix="product_formData.is_percent === true ? ' %' : ''"
                     :class="classes ? '' : ''"
-                    @change="calculateDiscount"
+                    @update:modelValue="calculateDiscount"
                     v-on="useListenerForm(product_formValidations, 'discount_value')"
                   />
                   <div class="absolute right-0 flex items-center rounded-lg border-none ring-0">
@@ -289,6 +350,7 @@ watch(product_formData, () => {
                       :options="['Rp', '%']"
                       class="border-none bg-transparent"
                       @update:modelValue="calculateDiscount"
+                      @change="calculateDiscount"
                     >
                       <template #dropdownicon>
                         <img :src="chevronDownSVG" alt="" />
@@ -342,7 +404,7 @@ watch(product_formData, () => {
                       v-model="product_formData.variants[index].name"
                       :name="`variants`"
                       class="border shadow-xs border-grayscale-30 rounded-lg w-full"
-                      :class="{ ...classes }"
+                      :class="classes"
                     />
                   </AppBaseFormGroup>
                 </div>
@@ -383,20 +445,32 @@ watch(product_formData, () => {
               </template>
             </PrimeVueButton>
           </div>
-          <div class="flex gap-4 mb-8">
-            <router-link to="/catalog/products">
+          <div class="flex items-center justify-between">
+            <div class="flex items-center gap-4">
+              <router-link to="/catalog/products">
+                <PrimeVueButton
+                  label="Cancel"
+                  class="text-xl w-48 py-2 border-2 border-primary cursor-pointer rounded-lg text-primary bg-transparent font-semibold"
+                  unstyled
+                />
+              </router-link>
               <PrimeVueButton
-                label="Cancel"
-                class="text-xl w-48 py-2 border-2 border-primary cursor-pointer rounded-lg text-primary bg-transparent font-semibold"
+                :label="'Edit Product'"
+                class="text-xl w-48 py-2 cursor-pointer border-2 border-primary rounded-lg text-white bg-primary font-semibold"
                 unstyled
+                type="submit"
               />
-            </router-link>
+            </div>
+
             <PrimeVueButton
-              :label="'Add Product'"
-              class="text-xl w-48 py-2 cursor-pointer border-2 border-primary rounded-lg text-white bg-primary font-semibold"
-              unstyled
-              type="submit"
-            />
+              label="Delete Product"
+              class="text-xl w-56 py-2 border-2 border-none cursor-pointer rounded-lg text-red-500 bg-transparent font-semibold"
+              @click="isDeleteOpen = true"
+            >
+              <template #icon>
+                <img :src="deleteSVG" alt="" />
+              </template>
+            </PrimeVueButton>
           </div>
         </div>
       </form>
@@ -405,7 +479,7 @@ watch(product_formData, () => {
       <template #container>
         <div class="w-[35rem] p-8">
           <div class="flex flex-col items-center gap-4 text-center">
-            <img :src="excludeSVG" alt="Delete icon" class="mx-auto" />
+            <span><img :src="excludeSVG" alt="" /></span>
             <h1 class="text-2xl font-semibold">Are you sure you want to leave this page?</h1>
             <p>Any changes you make to the data will be lost if you leave this page without saving</p>
             <div class="flex items-center justify-between gap-4">
@@ -421,6 +495,63 @@ watch(product_formData, () => {
                 @click="cancelLeave"
                 >Cancel</PrimeVueButton
               >
+            </div>
+          </div>
+        </div>
+      </template>
+    </PrimeVueDialog>
+
+    <PrimeVueDialog v-model:visible="isDeleteOpen" modal header="">
+      <template #container>
+        <div class="w-[35rem] p-8">
+          <div class="flex flex-col items-center gap-4 text-center">
+            <img :src="deletePolygonSVG" alt="Delete icon" class="mx-auto" />
+            <h1 class="text-2xl font-semibold">Are you sure you want to delete this product?</h1>
+            <p>This action cannot be undone, and the product will be removed from catalog</p>
+            <div class="flex items-center justify-between gap-4">
+              <PrimeVueButton
+                class="text-lg w-56 text-red-500 bg-transparent border-none"
+                variant="outlined"
+                label="Delete Product"
+                severity="danger"
+                @click="
+                  handleDelete();
+                  isDeleteOpen = false;
+                "
+              >
+                <template #icon>
+                  <img :src="deleteSVG" alt="" />
+                </template>
+              </PrimeVueButton>
+              <PrimeVueButton class="w-56 text-lg bg-primary border-primary" @click="isDeleteOpen = false"
+                >Cancel
+              </PrimeVueButton>
+            </div>
+          </div>
+        </div>
+      </template>
+    </PrimeVueDialog>
+
+    <PrimeVueDialog :visible="isUpdateModal" modal header="">
+      <template #container>
+        <div class="w-[35rem] p-8">
+          <div class="flex flex-col items-center gap-4 text-center">
+            <span><img :src="confirmationSVG" alt="" /></span>
+            <h1 class="text-2xl font-semibold">Are you sure want to update this product item?</h1>
+            <p>The update will affect the product items in the catalog</p>
+            <div class="flex items-center justify-between gap-4">
+              <PrimeVueButton
+                variant="text"
+                class="w-56 text-lg border-2 border-primary text-primary font-semibold"
+                @click="cancelUpdate"
+                >Cancel</PrimeVueButton
+              >
+              <PrimeVueButton
+                class="text-xl w-56 py-2 cursor-pointer border-2 border-primary rounded-lg text-white bg-primary font-semibold"
+                unstyled
+                label="Yes, I'm Sure"
+                @click="confirmUpdate"
+              />
             </div>
           </div>
         </div>
